+++ main.go
@@ -3,62 +3,68 @@ package main
 import (
 	"database/sql"
 	"fmt"
-	"net/http"
 	"os"
+	"url-shortener/internal/app/handlers"
+	"url-shortener/internal/config"
+	"url-shortener/internal/infrastructure/database"
+	"url-shortener/internal/infrastructure/database/migrations"
+	"url-shortener/internal/infrastructure/http"
 
 	"github.com/joho/godotenv"
-	"github.com/labstack/echo/v4"
-	"url-shortener/db"
 )
-
-// NewMySQLConnector creates a new MySQLConnector with the provided parameters.
-func NewMySQLConnector() *db.MySQLConnector {
-	return &db.MySQLConnector{
-		Username: os.Getenv("DB_USERNAME"),
-		Password: os.Getenv("DB_PASSWORD"),
-		Host:     os.Getenv("DB_HOST"),
-		Port:     os.Getenv("DB_PORT"),
-		DBName:   os.Getenv("DB_NAME"),
-	}
-}
 
 func main() {
 	// Load environment variables from .env file
 	err := godotenv.Load()
 	if err != nil {
-		fmt.Println("Error loading .env file:", err)
+		fmt.Println("[MAIN] Error loading .env file:", err)
 		return
 	}
 
-	// Create MySQLConnector with environment variables
-	connector := NewMySQLConnector()
+	// Create DBConnector with environment variables
+	connector := config.NewDBConnector()
+	dbDriver := "mysql"
 
 	// Connect to the database
-	database, err := db.ConnectToDB(connector)
+	db, err := database.ConnectToDB(connector, dbDriver)
 	if err != nil {
-		fmt.Println("Error connecting to database:", err)
+		fmt.Println("[MAIN] Error connecting to database:", err)
 		return
 	}
+
 	defer func(db *sql.DB) {
 		err := db.Close()
 		if err != nil {
-			fmt.Println("Error closing database connection:", err)
+			fmt.Println("[MAIN] Error closing database connection:", err)
 		}
-	}(database)
+	}(db)
 
-	fmt.Println("Connected to database")
+	fmt.Println("[DATABASE] Connected to database")
 
-	// Initialize Echo
-	e := echo.New()
-
-	// Define routes
-	e.GET("/", func(c echo.Context) error {
-		return c.String(http.StatusOK, "Hello, World!")
-	})
-
-	// Start the Echo server
-	err = e.Start(":1323")
+	// Run migrations
+	migrationsDir := os.Getenv("MIGRATIONS_DIR")
+	err = migrations.RunMigrations(db, migrationsDir)
 	if err != nil {
-		fmt.Println("Error starting Echo server:", err)
+		fmt.Println("[MAIN] Error running migrations:", err)
+		return
 	}
+
+	// Create auth handler
+	userHandler, err := handlers.InitializeUserHandlers(db)
+
+	if err != nil {
+		fmt.Println("[MAIN] Error initializing auth handlers:", err)
+		return
+	}
+
+	host := os.Getenv("HOST")
+	port := os.Getenv("PORT")
+	server := http.NewServer(host, port, userHandler)
+
+	err = server.Start()
+	if err != nil {
+		return
+	}
+
+	fmt.Println("[SERVER] Server started at", host+":"+port)
 }