+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 # Changelog
 All significant updates to this project will be meticulously documented in this log.
 
+## 0.1.0 - 08/03/2024
+
+### Changed
+
+- **Versioning:** Updated the versioning system to a stable release.
+
+## 0.0.0.alpha.4 - 08/03/2024
+
+### Added
+
+- **User Authentication:** Added user authentication to the project.
+
+- **JWT Library:** Added the JWT library to the project.
+
+- **Auth Repository:** Added an auth repository to handle user authentication.
+
+- **Auth Service:** Added an auth service to handle user authentication.
+
+- **Auth Handler:** Added an auth handler to handle user authentication.
+
+## 0.0.0.alpha.3 - 08/03/2024
+
+### Added
+
+- **User Model:** Added a user model to the project.
+
+- **Migration:** Added a migration to create the users table.
+
+- **Echo Server:** Added a new route to the echo server to handle user registration.
+
+- **Echo Library:** Added the echo library to the project.
+
+- **Added New Instructions For Makefile:** Added new instructions to the Makefile to handle the new changes.
+
+### Changed
+
+- **Database:** Updated the database schema to include the users table.
+
+- **README:** Updated the instructions in the README file.
+
 ## 0.0.0.alpha.2 - 07/03/2024
 
 ### Added
@@ -14,4 +54,3 @@ All significant updates to this project will be meticulously documented in this 
 - **Changelog File:** Added a new changelog file to keep track of the changes made to the project.
 
 - **Versioning:** Added a versioning system to the project.
-